Refactor the blocks
