--- wagtailio/newsletter/templates/newsletter/blocks/content_block.html
@@ -1,2 +1,0 @@
-{% load wagtailcore_tags %}
-{{ value.text|richtext }}
