+++ wagtailio/newsletter/templates/newsletter/blocks/heading_block.html
@@ -1,5 +1,1 @@
-{% if value.size == 'h1' %}
-    <h1>{{ value.heading }}</h1>
-{% else %}
-    <h2>{{ value.heading }}</h2>
-{% endif %}
+<{{ value.size }}>{{ value.heading }}</{{ value.size }}>
