--- wagtailio/newsletter/templates/newsletter/blocks/image_block.html
@@ -1,8 +1,0 @@
-{% load wagtailimages_tags %}
-
-<figure>
-    {% image value.image width-564 alt=value.alt_text %}
-    {% if value.caption %}
-        <figcaption>{{ value.caption }}</figcaption>
-    {% endif %}
-</figure>
