--- wagtailio/newsletter/templates/newsletter/blocks/package_block.html
@@ -1,1 +1,0 @@
-<p><a href="{{ value.url }}" target="_blank">{{ value.name }} v{{ value.version }}</a></p>
