--- wagtailio/newsletter/templates/newsletter/blocks/package_list_block.html
@@ -1,6 +1,0 @@
-<div class="package-list">
-    <h2>{{ value.title }}</h2>
-    {% for package in value.packages %}
-        {% include "newsletter/blocks/package_block.html" with value=package %}
-    {% endfor %}
-</div>
